backend/api_base_public/requirements.txt
# requirements.txt

annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
colorama==0.4.6
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysql-connector-python==9.3.0
pyasn1==0.6.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.19
passlib==1.7.4
bcrypt==4.1.3
httpx==0.28.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1

# === Comic Engine Dependencies (từ THUC_TAP2) ===
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# === AI/ML (Optional - comment out nếu không cần) ===
# ultralytics>=8.0.0       # YOLO - phân loại nhân vật (nặng ~100MB)
# transformers>=4.30.0     # CLIP - phân loại cảnh (nặng)
# torch>=2.0.0             # PyTorch (nặng)
# face-recognition>=1.3.0  # Nhận diện khuôn mặt (cần dlib)
